There was no overflow error; my mistake!

+++ _samples/2017/07/equals/src/main/java/Comparison.java
@@ -42,7 +42,7 @@ class Comparison<T extends Digitizable> {
             if (idx <= right.length) {
                 rht = right[max - idx];
             }
-            result = (int) lft - (int) rht;
+            result = lft - rht;
             if (result != 0) {
                 break;
             }
